[FEATURE][SWE-1792]fixed PR issues
