+++ src/domains/Booking/__constants__/bookingCancelTimeLeft.js
@@ -1,0 +1,3 @@
+const BOOKINGS_CANCEL_TIME_LIMIT = 48
+
+export default BOOKINGS_CANCEL_TIME_LIMIT
