+++ src/domains/Booking/hooks/get/index.js
@@ -2,3 +2,4 @@ export { default as useGetNextBookingData } from './useGetNextBookingData'
 export { default as useBookings } from './useBookings'
 export { default as useGetBookingsCounter } from './useGetBookingsCounter'
 export { default as useGetBookingData } from './useGetBookingData'
+export { default as useGetBookingDate } from './useGetBookingDate'
